fix and enable md5 for reference.

+++ md5.h
@@ -4,8 +4,8 @@
  */
 typedef struct
 {
-    unsigned long state[4];     /*!< intermediate digest state  */
-    unsigned long total[2];     /*!< number of bytes processed  */
+    uint32_t state[4];     /*!< intermediate digest state  */
+    uint32_t total[2];     /*!< number of bytes processed  */
     unsigned char buffer[64];   /*!< data block being processed */
 
     unsigned char ipad[64];     /*!< HMAC: inner padding        */
